move to util/types
