+++ static/app/utils/types/flatten.tsx
