+++ static/app/views/insights/common/components/tableCells/starredSegmentCell.tsx
@@ -3,8 +3,8 @@ import {IconStar} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import type {EventsMetaType} from 'sentry/utils/discover/eventView';
 import {FlexContainer} from 'sentry/utils/discover/styles';
-import type {Flatten} from 'sentry/utils/flatten';
 import {useQueryClient} from 'sentry/utils/queryClient';
+import type {Flatten} from 'sentry/utils/types/flatten';
 import useProjects from 'sentry/utils/useProjects';
 import {useStarredSegment} from 'sentry/views/insights/common/utils/useStarredSegment';
 import type {EAPSpanResponse} from 'sentry/views/insights/types';
